rename config page
