+++ Dockerfile
@@ -9,5 +9,7 @@ RUN pip install --upgrade --no-cache-dir \
 
 ADD ./trans_rss /app/trans_rss
 
+EXPOSE 80
+
 CMD ["uvicorn", "trans_rss:app", "--host", "0.0.0.0", "--port", "80"]
