+++ build.sh
@@ -1,2 +1,2 @@
 version=$(cat trans_rss/version)
-podman build . -t docker.io/liyihc/trans-rss:$version -t docker.io/liyihc/trans-rss:latest
+docker build . -t docker.io/liyihc/trans-rss:$version -t docker.io/liyihc/trans-rss:latest
