+++ trans_rss/web/__init__.py
@@ -2,7 +2,7 @@ from pywebio.platform.fastapi import webio_routes
 
 from trans_rss.config import config
 
-from .config import config_page
+from .config_page import config_page
 from .logs import log_page
 from .manage import manage_subscribe_page
 from .subscribe import sub_list_page, subscribe_page
@@ -19,5 +19,5 @@ routes = webio_routes(
         "subscribe-manage": manage_subscribe_page,
         "webhook-type": webhook_type_page,
         "subscribe-type": subscribe_type_page
-    }, cdn=True if config.cdn else "static"
+    }, cdn=True if config.cdn else "static" # type: ignore
 )
